Made cleaner names for model functions
Made cleaner names for model functions. For example:
falco.models.model_full() --> falco.model.full()
falco.models.model_compact() --> falco.model.compact()
falco.models.model_Jacobian() --> falco.model.jacobian()
falco.models.model_Jacobian_LC() --> falco.model.jacobians.LC()
