+++ falco/__init__.py
@@ -8,7 +8,7 @@ from .dms import *
 from .setups import *
 from .propcustom import *
 from .zernikes import *
-from . import models
+from . import model
 from ._globals import INFLUENCE_XINETICS
 from ._globals import INFLUENCE_BMC_KILO
 from ._globals import INFLUENCE_BMC_2K
